app/requirements.txt
apiclient==1.0.4
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
flake8==4.0.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.1
idna==3.3
importlib-metadata==4.2.0
isort==5.10.1
mccabe==0.6.1
oauth2client==4.1.3
psycopg2-binary==2.9.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rsa==4.0
six==1.12.0
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritemplate==3.0.0
urllib3==1.26.9
zipp==3.8.0
